Changed process_rsmas logic flow
`process_rsmas` now either submits `process_rsmas` as a job file via
`create_batch.submit_script()` or runs the remainder of the routine in
a serial fashion, but not both. This should guarantee that, with the
removal of the `sys.exit()` call from `create_batch`, the main body of
`process_rsmas` is not run twice.